+++ acme/services/Onboarding.py
@@ -338,8 +338,7 @@ allowedCSROriginators=id-in,id-mn,id-asn
 
 		cnfRegular = \
 """
-[scripting]
-scriptDirectories=${cse:resourcesPath}/utilities
+
 """
 
 
@@ -354,9 +353,6 @@ enable=true
 [http]
 enableUpperTesterEndpoint=true
 enableStructureEndpoint=true
-
-[scripting]
-scriptDirectories=${cse:resourcesPath}/utilities
 """
 
 		cnfIntroduction = \
@@ -368,7 +364,7 @@ startWithTUI=true
 enable=true
 
 [scripting]
-scriptDirectories=${cse:resourcesPath}/demoLightbulb,${cse:resourcesPath}/demoDocumentationTutorials,${cse:resourcesPath}/utilities
+scriptDirectories=${cse:resourcesPath}/demoLightbulb,${cse:resourcesPath}/demoDocumentationTutorials
 """
 
 		cnfHeadless = \